+++ common/misc.h
@@ -3,7 +3,7 @@
 
 #include "common/types.h"
 
-bool HttpGetFileSize(const std::string& url, __out size_t& file_size);
+bool HttpGetFileSize(const std::string& url, __out unsigned long long& file_size);
 
 bool HttpReadFileToBuffer(const std::string& url, void *buf, size_t size, __out size_t& read_size);
